Use an integer type for sNumLivingContexts that won't overflow when counting objects.

+++ dom/canvas/CanvasRenderingContext2D.cpp
@@ -923,7 +923,7 @@ NS_INTERFACE_MAP_END
 
 
 // Initialize our static variables.
-uint32_t CanvasRenderingContext2D::sNumLivingContexts = 0;
+uintptr_t CanvasRenderingContext2D::sNumLivingContexts = 0;
 DrawTarget* CanvasRenderingContext2D::sErrorTarget = nullptr;
